medconsult/requirements.txt
torch
transformers
accelerate
gradio
Pillow
pytest
chromadb
google-generativeai
